pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-tool-hub"
version = "0.1.0"
description = "A FastMCP hub server that manages and dispatches tools across downstream MCP servers."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastmcp>=0.2.0",
  "mcp>=1.0.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
mcp-tool-hub = "tool_hub.hub_server:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

